+++ R/dbcAlert.R
@@ -1,9 +1,9 @@
 # AUTO GENERATED FILE - DO NOT EDIT
 
 #' @export
-dbcAlert <- function(children=NULL, id=NULL, className=NULL, class_name=NULL, color=NULL, dismissable=NULL, duration=NULL, fade=NULL, is_open=NULL, key=NULL, loading_state=NULL, style=NULL) {
+dbcAlert <- function(children=NULL, id=NULL, className=NULL, class_name=NULL, color=NULL, dismissable=NULL, duration=NULL, fade=NULL, is_open=NULL, key=NULL, loading_state=NULL, persisted_props=NULL, persistence=NULL, persistence_type=NULL, style=NULL) {
     
-    props <- list(children=children, id=id, className=className, class_name=class_name, color=color, dismissable=dismissable, duration=duration, fade=fade, is_open=is_open, key=key, loading_state=loading_state, style=style)
+    props <- list(children=children, id=id, className=className, class_name=class_name, color=color, dismissable=dismissable, duration=duration, fade=fade, is_open=is_open, key=key, loading_state=loading_state, persisted_props=persisted_props, persistence=persistence, persistence_type=persistence_type, style=style)
     if (length(props) > 0) {
         props <- props[!vapply(props, is.null, logical(1))]
     }
@@ -11,7 +11,7 @@ dbcAlert <- function(children=NULL, id=NULL, className=NULL, class_name=NULL, co
         props = props,
         type = 'Alert',
         namespace = 'dash_bootstrap_components',
-        propNames = c('children', 'id', 'className', 'class_name', 'color', 'dismissable', 'duration', 'fade', 'is_open', 'key', 'loading_state', 'style'),
+        propNames = c('children', 'id', 'className', 'class_name', 'color', 'dismissable', 'duration', 'fade', 'is_open', 'key', 'loading_state', 'persisted_props', 'persistence', 'persistence_type', 'style'),
         package = 'dashBootstrapComponents'
         )