+++ R/dbcCarousel.R
@@ -1,9 +1,9 @@
 # AUTO GENERATED FILE - DO NOT EDIT
 
 #' @export
-dbcCarousel <- function(id=NULL, active_index=NULL, className=NULL, class_name=NULL, controls=NULL, indicators=NULL, interval=NULL, items=NULL, loading_state=NULL, ride=NULL, slide=NULL, style=NULL, variant=NULL) {
+dbcCarousel <- function(id=NULL, active_index=NULL, className=NULL, class_name=NULL, controls=NULL, indicators=NULL, interval=NULL, items=NULL, loading_state=NULL, persisted_props=NULL, persistence=NULL, persistence_type=NULL, ride=NULL, slide=NULL, style=NULL, variant=NULL) {
     
-    props <- list(id=id, active_index=active_index, className=className, class_name=class_name, controls=controls, indicators=indicators, interval=interval, items=items, loading_state=loading_state, ride=ride, slide=slide, style=style, variant=variant)
+    props <- list(id=id, active_index=active_index, className=className, class_name=class_name, controls=controls, indicators=indicators, interval=interval, items=items, loading_state=loading_state, persisted_props=persisted_props, persistence=persistence, persistence_type=persistence_type, ride=ride, slide=slide, style=style, variant=variant)
     if (length(props) > 0) {
         props <- props[!vapply(props, is.null, logical(1))]
     }
@@ -11,7 +11,7 @@ dbcCarousel <- function(id=NULL, active_index=NULL, className=NULL, class_name=N
         props = props,
         type = 'Carousel',
         namespace = 'dash_bootstrap_components',
-        propNames = c('id', 'active_index', 'className', 'class_name', 'controls', 'indicators', 'interval', 'items', 'loading_state', 'ride', 'slide', 'style', 'variant'),
+        propNames = c('id', 'active_index', 'className', 'class_name', 'controls', 'indicators', 'interval', 'items', 'loading_state', 'persisted_props', 'persistence', 'persistence_type', 'ride', 'slide', 'style', 'variant'),
         package = 'dashBootstrapComponents'
         )