+++ R/dbcPopover.R
@@ -1,9 +1,9 @@
 # AUTO GENERATED FILE - DO NOT EDIT
 
 #' @export
-dbcPopover <- function(children=NULL, id=NULL, autohide=NULL, body=NULL, className=NULL, class_name=NULL, delay=NULL, flip=NULL, hide_arrow=NULL, innerClassName=NULL, inner_class_name=NULL, is_open=NULL, key=NULL, loading_state=NULL, offset=NULL, placement=NULL, style=NULL, target=NULL, trigger=NULL) {
+dbcPopover <- function(children=NULL, id=NULL, autohide=NULL, body=NULL, className=NULL, class_name=NULL, delay=NULL, flip=NULL, hide_arrow=NULL, innerClassName=NULL, inner_class_name=NULL, is_open=NULL, key=NULL, loading_state=NULL, offset=NULL, persisted_props=NULL, persistence=NULL, persistence_type=NULL, placement=NULL, style=NULL, target=NULL, trigger=NULL) {
     
-    props <- list(children=children, id=id, autohide=autohide, body=body, className=className, class_name=class_name, delay=delay, flip=flip, hide_arrow=hide_arrow, innerClassName=innerClassName, inner_class_name=inner_class_name, is_open=is_open, key=key, loading_state=loading_state, offset=offset, placement=placement, style=style, target=target, trigger=trigger)
+    props <- list(children=children, id=id, autohide=autohide, body=body, className=className, class_name=class_name, delay=delay, flip=flip, hide_arrow=hide_arrow, innerClassName=innerClassName, inner_class_name=inner_class_name, is_open=is_open, key=key, loading_state=loading_state, offset=offset, persisted_props=persisted_props, persistence=persistence, persistence_type=persistence_type, placement=placement, style=style, target=target, trigger=trigger)
     if (length(props) > 0) {
         props <- props[!vapply(props, is.null, logical(1))]
     }
@@ -11,7 +11,7 @@ dbcPopover <- function(children=NULL, id=NULL, autohide=NULL, body=NULL, classNa
         props = props,
         type = 'Popover',
         namespace = 'dash_bootstrap_components',
-        propNames = c('children', 'id', 'autohide', 'body', 'className', 'class_name', 'delay', 'flip', 'hide_arrow', 'innerClassName', 'inner_class_name', 'is_open', 'key', 'loading_state', 'offset', 'placement', 'style', 'target', 'trigger'),
+        propNames = c('children', 'id', 'autohide', 'body', 'className', 'class_name', 'delay', 'flip', 'hide_arrow', 'innerClassName', 'inner_class_name', 'is_open', 'key', 'loading_state', 'offset', 'persisted_props', 'persistence', 'persistence_type', 'placement', 'style', 'target', 'trigger'),
         package = 'dashBootstrapComponents'
         )